refactor(web, api): rename user folder id to userId
