+++ src/pages/api/users/[userId]/chat/[chatId]/completions.ts
@@ -2,7 +2,7 @@ import { type NextApiRequest, type NextApiResponse } from 'next';
 import { z } from 'zod';
 import { getCompletions } from '@/lib/openai';
 import prisma from '@/lib/prisma';
-import { getOneChat } from '@/pages/api/users/[id]/chat/[chatId]';
+import { getOneChat } from '@/pages/api/users/[userId]/chat/[chatId]';
 
 export const completionBodySchema = z.object({
   model: z.string(),
@@ -15,7 +15,7 @@ export const completionBodySchema = z.object({
 });
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
-  const { id: userId, chatId } = req.query;
+  const { userId, chatId } = req.query;
 
   if (!userId || !chatId) {
     return res.status(400).json({ error: 'Bad request' });
