+++ src/pages/api/users/[userId]/chat/[chatId]/prompt.ts
@@ -1,7 +1,7 @@
 import { type NextApiRequest, type NextApiResponse } from 'next';
 import { z } from 'zod';
 import prisma from '@/lib/prisma';
-import { getOneChat } from '@/pages/api/users/[id]/chat/[chatId]';
+import { getOneChat } from '@/pages/api/users/[userId]/chat/[chatId]';
 
 export const promptBodySchema = z.object({
   content: z.string().max(191),
@@ -11,7 +11,7 @@ export const promptBodySchema = z.object({
 });
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
-  const { id: userId, chatId } = req.query;
+  const { userId, chatId } = req.query;
 
   if (!userId || !chatId) {
     return res.status(400).json({ error: 'Bad request' });
