+++ src/pages/api/users/[userId]/chat/[chatId]/regenerate.ts
@@ -2,11 +2,11 @@ import { type NextApiRequest, type NextApiResponse } from 'next';
 import { z } from 'zod';
 import { getCompletions } from '@/lib/openai';
 import prisma from '@/lib/prisma';
-import { getOneChat } from '@/pages/api/users/[id]/chat/[chatId]';
-import { completionBodySchema } from '@/pages/api/users/[id]/chat/[chatId]/completions';
+import { getOneChat } from '@/pages/api/users/[userId]/chat/[chatId]';
+import { completionBodySchema } from '@/pages/api/users/[userId]/chat/[chatId]/completions';
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
-  const { id: userId, chatId } = req.query;
+  const { userId, chatId } = req.query;
 
   if (!userId || !chatId) {
     return res.status(400).json({ error: 'Bad request' });
